Increase precision of EFF to one's digit
We now calculate the full 9 bits of EFF, rounding down. Because we still
truncate the one's bit of lines if lines is large, there is still source
of error. If lines > 255 and is odd, the raw calculation will be too
high with a worst-case of 1.17. This worst-case occurs if the actual EFF
is near 300 and lines is near 256. It does not appear possible to abuse
this to obtain an EFF > 300.

Note that truncating the one's bit of lvl0 score has impact on the
accuracy, since all scores are divisible by 4.