mod: Add playerid display
On the level menu screen, you can press select to swap through player
ids in the top-right.

This is currently implemented in two fashions: background and sprite.
It's not clear which one is better. Defaults to background. Set
PLAYERID_SPRITE=1 (e.g., at top of tetris-mod.s) to enable sprite
version.