tetris-PRG.info: Name more local labels
Using @ makes it much easier to come up with a name and makes it clear
we are still within a "function." A mod may need one of these cheap
local labels later, in which case we'll have to figure out how we want
to handle it. But for now, it improves readability enough to be worth
it.