Add vertical chart for EFF log
Horizontal chart also implemented, but while easier to implement seems
harder to read.

The vertical chart is techincally still incomplete, as it will
easily exceed the 8 sprites per scanline limit. There are two approaches
to fix that, and both already have the necessary tiles in taus.chr. One
is to calculate the difference between two bars and use a single sprite
for two bars when they overlap. The alternative is to use a second
sprite with the endcap at the bottom, and obscure any bar above the
endcap. This approach, however, would require modifying the palette and
is also incompatible with the gridlines.

Even in its final form, there is a chance the 8 sprite limit could be
visible, but it would take a 190+ line game with consistent EFF. I'm
just accepting that will be rare enough to let a glitch occur in that
case.

The rendering approach could still be improved. Currently there is a
black screen flash when rendering. It also interacts strangly with the
curtain. Probably should render to playfield and then set
player1_vramRow to render it to the screen, but such improvements are
for later work.