twoplayer: Flicker when 8-sprites-per-scanline exceeded
Flickering seems fairest, as each player can still continue like normal.
Didn't bother messing with player2's next pice, as it seems the problem
will rarely be triggered given how low in the playfield it is and the
next piece is the thing that won't display, which is better than the
current piece.

The pathological case for triggering this is when there is no input for
either player, as then the falling and pieces are synchronized. I
seriously believe it may be a while before anyone notices during normal
play.