towplayer: Optimize level mod10 for higher levels
This is 3 cycles slower for level <10. But it saves 6 cycles each loop
iteration, so is 4 cycles faster for levels in the 10s and 10 cycles
faster for levels in the 20s.

This does mean that lower levels no longer finish OAMDMA during vsync.
But the higher levels are important and we want to reduce the worst-case
costs.