twoplayer: Flicker between more pieces for 8 sprite limit
This makes pieces brighter when flickering, although it makes more of
the screen flicker.

This only handles 10 sprites per scanline. While 12 total is possible,
it requires two horizontal Is which is highly unlikely in legitimate
play. It'd be easy to make it work for 12 sprites, but there doesn't
seem to be much point.