twoplayer: Add controls for Start-less Famicom P2
Since P2 lacks a Start button on Famicom, we now let P2 press A to join.
We also let P1 start the game with Start. If P2 is holding A when P1
presses start they will have +10 added to the level. The old controls
are still supported.

No detection of the Famicom is included. These keys are always allowed.
Also, the on-screen text still says "PRESS START!" even though "A" is
necessary on Famicom.

Fixes #6