twoplayer: Use ips for fastlegal for PAL
The PAL rom uses $CC instead of NTSC's $FF for the number of frames to
wait for the legal screen. This prevents the twoplayer patch from
applying. Using the fastlegal IPS side-steps the problem and fixes
building with PAL=1. The results are still very broken, but at least it
builds.