twoplayer: Add compatibility with TetrisControllerInputDisplay
TetrisControllerInputDisplay uses the same ROM area to store its code as
twoplayer. Just shift twoplayer a bit so the two no longer collide.

V1.0 of the display mod zeroed out a ROM area it didn't use; for that
version you must apply the display mod before twoplayer. V1.1 fixes that
so the two can be applied in any order.

Fixes #19