twoplayer: Use unreferenced_data4 for tournament mode
This approach of carving up the original ROM isn't scaling well.
Overwrite segments in cc65 would probably work nicely, but they aren't
quite working yet (https://github.com/cc65/cc65/issues/1366). Even after
it is fixed and released, it is unclear if I want to depend on that new
of a cc65.

I've considered using the IPS approach more, but that is too much work
and risky for the moment. Let's take the simple approach and get
-DTOURNAMENT_MODE compiling again.