save-highscores: Check for uninitialized SRAM; use default high scores
High scores are now initialized to the default high scores instead of
zeros. The magic is moved to the SRAM to know if we've already
initialized SRAM, and we memset all of SRAM to zero in case we want to
save something else later without clearing high scores.

Instead of placing our hack's code in defaultHighScoreTable, it is now
in the cold boot detection and logic.

Fixes #25