Fix 7bag and other jsr()-using tests
The strangeness of jsr() was fixed such that now we can have the CPU to
ourselves for pure-computation tests. No more weird double-waitbefore
and waiting between each iteration. More importantly, it is much more
reliable. 7bag was suffering because of the earlier jsr() strangeness
corrupting results and wasn't stable across FCEUX versions.